// src/app/(modulos)/operaciones/models/models.ts
export interface Supplier {
  code: string;
  name: string;
  ruc: string;
  isActive: string;
  storageCode: string;
  initials: string;
  emails: string[];
  addresses: Record<string, string>;
}

export interface FabricType {
  id: number;
  value: string;
}

export interface FabricRecipe {
  yarnId: string; 
  proportion: number; 
  numPlies: number; 
  gauge: number; 
  diameter: number; 
  stitchLength: number;
}

export interface Fabric {
  id: string;
  fabricTypeId: number;
  density: number;
  width: number;
  colorId: string | null;
  structurePattern: string;
  description: string;
  isActive: boolean;
  recipe: FabricRecipe[];
  fabricType: {
    id: number;
    value: string;
  };
}

// Para la respuesta de GET /operations/v1/fabrics
export interface FabricResponse {
  fabrics: Fabric[];
}

export interface WeavingServiceEntry {
  entryNumber: string;
  period: number;
  creationDate: string;
  creationTime: string;
  supplierCode: string;
  statusFlag: string;
  supplierPoCorrelative?: string;
  supplierPoSeries?: string;
  documentNote?: string;
  fecgf?: string;
  userId?: string;
  detail?: Detail[];
}

export interface Detail {
  itemNumber: number;
  fabridId: string;
  mecsaWeight: number;
  statusFlag: string;
  serviceOrderId: string;
  detailCard: DetailCard[];
  guideNetWeight: number;
  rollCount: number;
  fabricType: string;
  tintColorId: string;
  tintSupplierId: string;
  tintSupplierColorId: string | null;
}

export interface DetailCard {
  id: string;
  fabricId: string;
  netWeight: number;
  tintSupplierId: string;
  tintColorId: string;
  yarnSupplierId: string;
  cardType: string;
  productId: string;
  statusFlag: string;
  documentNumber: string;
  exitNumber: string;
  serviceOrderId: string;
  supplierWeavingTej: string;
  suppliersYarn: string[];
  serviceOrders: string[];
}